+++ src/mefit/Forms/frmEfiRom.cs
@@ -1422,8 +1422,8 @@ namespace Mac_EFI_Toolkit.Forms
         {
             var buttons = new[]
             {
-                new { Button = cmdClose, Font = Program.FontSegMdl2Regular12, Text = Program.GLYPH_EXIT_CROSS },
-                new { Button = cmdOpenInExplorer, Font = Program.FontSegMdl2Regular12, Text = Program.GLYPH_FILE_EXPLORER },
+                new { Button = cmdClose, Font = Program.FontSegMdl2Regular12, Text = Program.MDL2_EXIT_CROSS },
+                new { Button = cmdOpenInExplorer, Font = Program.FontSegMdl2Regular12, Text = Program.MDL2_FILE_EXPLORER },
             };
 
             foreach (var buttonData in buttons)
@@ -1436,10 +1436,10 @@ namespace Mac_EFI_Toolkit.Forms
         private void SetLabelFontAndGlyph()
         {
             lblLzma.Font = Program.FontSegMdl2Regular10;
-            lblLzma.Text = Program.GLYPH_REPORT;
+            lblLzma.Text = Program.MDL2_REPORT;
 
             lblFmmEmail.Font = Program.FontSegMdl2Regular10;
-            lblFmmEmail.Text = Program.GLYPH_ACCOUNT;
+            lblFmmEmail.Text = Program.MDL2_ACCOUNT;
         }
 
         private void SetTipHandlers()
@@ -1555,7 +1555,7 @@ namespace Mac_EFI_Toolkit.Forms
         private void UpdateWindowTitle()
         {
             this.Text = EFIROM.FileInfoData.FileNameExt;
-            lblTitle.Text = $"{APPSTRINGS.EFIROM} {Program.GLYPH_RIGHT_ARROW} {EFIROM.FileInfoData.FileNameExt}";
+            lblTitle.Text = $"{APPSTRINGS.EFIROM} {Program.MDL2_RIGHT_ARROW} {EFIROM.FileInfoData.FileNameExt}";
         }
         #endregion